KafkaDocker/producer/requirements.txt
kafka-python
jason